<fix>[vm]: update APIGetVmsCapabilitiesMsg RestRequest definition
APIImpact

Resolves: ZSTAC-62246
